src/agents/search_agent.py
"""
Search Agent for real-time ESG regulation updates and research
"""
import asyncio
import aiohttp
import json
from datetime import datetime, timedelta
from typing import Dict, Any, List, Optional
from bs4 import BeautifulSoup
import feedparser

from .base_agent import BaseAgent


class SearchAgent(BaseAgent):
    """Agent responsible for searching and gathering ESG-related information"""
    
    def __init__(self, config: Dict[str, Any] = None):
        super().__init__(
            name="SearchAgent",
            description="Searches for real-time ESG regulations, standards, and best practices",
            config=config
        )
        self.search_sources = [
            {
                "name": "SEC ESG Updates",
                "url": "https://www.sec.gov/news/pressreleases",
                "type": "regulatory"
            },
            {
                "name": "GRI Standards",
                "url": "https://www.globalreporting.org/news/",
                "type": "standards"
            },
            {
                "name": "SASB Updates",
                "url": "https://www.sasb.org/news/",
                "type": "standards"
            }
        ]
        self.cache = {}
        self.cache_duration = timedelta(hours=6)
    
    async def execute(self, task: Dict[str, Any]) -> Dict[str, Any]:
        """Execute search task"""
        task_type = task.get("type", "general_search")
        
        if task_type == "regulation_search":
            return await self.search_regulations(task.get("query", ""))
        elif task_type == "standards_update":
            return await self.get_standards_updates()
        elif task_type == "industry_benchmarks":
            return await self.get_industry_benchmarks(task.get("industry", ""))
        else:
            return await self.general_esg_search(task.get("query", ""))
    
    async def search_regulations(self, query: str) -> Dict[str, Any]:
        """Search for ESG regulations"""
        # Mock implementation - in production, this would use real APIs
        await asyncio.sleep(1)  # Simulate API call
        
        mock_regulations = [
            {
                "title": "SEC Climate Disclosure Rules Update",
                "source": "SEC",
                "date": "2024-01-15",
                "summary": "New requirements for climate-related disclosures in annual reports",
                "impact": "High",
                "url": "https://www.sec.gov/rules/proposed/2022/33-11042.pdf"
            },
            {
                "title": "EU Taxonomy Regulation Amendment",
                "source": "European Commission",
                "date": "2024-01-10",
                "summary": "Updated criteria for sustainable economic activities",
                "impact": "Medium",
                "url": "https://ec.europa.eu/info/business-economy-euro/banking-and-finance/sustainable-finance/eu-taxonomy-sustainable-activities_en"
            },
            {
                "title": "CSRD Implementation Guidelines",
                "source": "EFRAG",
                "date": "2024-01-05",
                "summary": "Corporate Sustainability Reporting Directive implementation details",
                "impact": "High",
                "url": "https://www.efrag.org/Activities/1895/Corporate-Sustainability-Reporting-Directive-CSRD"
            }
        ]
        
        # Filter based on query if provided
        if query:
            filtered_regulations = [
                reg for reg in mock_regulations
                if query.lower() in reg["title"].lower() or query.lower() in reg["summary"].lower()
            ]
        else:
            filtered_regulations = mock_regulations
        
        return {
            "regulations": filtered_regulations,
            "total_found": len(filtered_regulations),
            "last_updated": datetime.now().isoformat(),
            "query": query
        }
    
    async def get_standards_updates(self) -> Dict[str, Any]:
        """Get latest ESG standards updates"""
        await asyncio.sleep(0.8)  # Simulate API call
        
        mock_standards = [
            {
                "standard": "GRI Universal Standards",
                "version": "2023",
                "update_date": "2023-12-01",
                "changes": [
                    "Enhanced materiality assessment requirements",
                    "New stakeholder engagement guidelines",
                    "Updated reporting principles"
                ],
                "effective_date": "2024-01-01"
            },
            {
                "standard": "SASB Standards",
                "version": "2023",
                "update_date": "2023-11-15",
                "changes": [
                    "Industry-specific metric updates",
                    "New disclosure topics for technology sector",
                    "Revised accounting metrics"
                ],
                "effective_date": "2024-01-01"
            },
            {
                "standard": "TCFD Recommendations",
                "version": "2023",
                "update_date": "2023-10-30",
                "changes": [
                    "Enhanced scenario analysis guidance",
                    "New metrics for transition risks",
                    "Updated governance recommendations"
                ],
                "effective_date": "2024-01-01"
            }
        ]
        
        return {
            "standards": mock_standards,
            "total_standards": len(mock_standards),
            "last_updated": datetime.now().isoformat()
        }
    
    async def get_industry_benchmarks(self, industry: str) -> Dict[str, Any]:
        """Get industry-specific ESG benchmarks"""
        await asyncio.sleep(1.2)  # Simulate API call
        
        # Mock industry benchmarks
        benchmarks = {
            "technology": {
                "carbon_intensity": {"average": 2.5, "top_quartile": 1.2, "unit": "tCO2e/revenue"},
                "diversity_ratio": {"average": 0.35, "top_quartile": 0.45, "unit": "ratio"},
                "board_independence": {"average": 0.75, "top_quartile": 0.85, "unit": "ratio"}
            },
            "manufacturing": {
                "carbon_intensity": {"average": 8.2, "top_quartile": 4.1, "unit": "tCO2e/revenue"},
                "waste_recycling": {"average": 0.65, "top_quartile": 0.85, "unit": "ratio"},
                "safety_incidents": {"average": 2.1, "top_quartile": 0.5, "unit": "per 100 employees"}
            },
            "financial": {
                "sustainable_finance": {"average": 0.25, "top_quartile": 0.45, "unit": "ratio"},
                "diversity_ratio": {"average": 0.40, "top_quartile": 0.55, "unit": "ratio"},
                "governance_score": {"average": 7.2, "top_quartile": 8.5, "unit": "score/10"}
            }
        }
        
        industry_data = benchmarks.get(industry.lower(), benchmarks["technology"])
        
        return {
            "industry": industry,
            "benchmarks": industry_data,
            "data_date": "2024-01-01",
            "sample_size": 150,
            "last_updated": datetime.now().isoformat()
        }
    
    async def general_esg_search(self, query: str) -> Dict[str, Any]:
        """Perform general ESG-related search"""
        await asyncio.sleep(0.6)  # Simulate API call
        
        mock_results = [
            {
                "title": "Best Practices in ESG Reporting",
                "source": "ESG Today",
                "date": "2024-01-20",
                "summary": "Comprehensive guide to effective ESG reporting strategies",
                "relevance": 0.95,
                "url": "https://example.com/esg-reporting-best-practices"
            },
            {
                "title": "Carbon Accounting Methodologies",
                "source": "CDP",
                "date": "2024-01-18",
                "summary": "Updated methodologies for measuring and reporting carbon emissions",
                "relevance": 0.88,
                "url": "https://example.com/carbon-accounting"
            },
            {
                "title": "Stakeholder Engagement in ESG",
                "source": "Sustainability Accounting Standards Board",
                "date": "2024-01-15",
                "summary": "Framework for effective stakeholder engagement in ESG initiatives",
                "relevance": 0.82,
                "url": "https://example.com/stakeholder-engagement"
            }
        ]
        
        # Filter based on query relevance
        if query:
            filtered_results = [
                result for result in mock_results
                if query.lower() in result["title"].lower() or query.lower() in result["summary"].lower()
            ]
        else:
            filtered_results = mock_results
        
        return {
            "results": filtered_results,
            "total_found": len(filtered_results),
            "query": query,
            "search_time": datetime.now().isoformat()
        }
    
    async def get_cached_result(self, cache_key: str) -> Optional[Dict[str, Any]]:
        """Get cached search result if still valid"""
        if cache_key in self.cache:
            cached_data = self.cache[cache_key]
            if datetime.now() - cached_data["timestamp"] < self.cache_duration:
                return cached_data["data"]
        return None
    
    def cache_result(self, cache_key: str, data: Dict[str, Any]):
        """Cache search result"""
        self.cache[cache_key] = {
            "data": data,
            "timestamp": datetime.now()
        }
    
    def clear_cache(self):
        """Clear all cached results"""
        self.cache.clear()