requirements.txt
# Core Streamlit and web framework
streamlit>=1.28.0

# Agentic AI framework
langchain>=0.1.0
langchain-community>=0.0.10
crewai>=0.28.0

# Mistral AI integration
mistralai>=0.4.0
langchain-mistralai>=0.1.0

# Data processing and analysis
pandas>=2.0.0
numpy>=1.24.0

# Visualization and charting
plotly>=5.17.0
matplotlib>=3.7.0
seaborn>=0.12.0

# PDF generation and reporting
reportlab>=4.0.0
pillow>=10.0.0

# Document processing and analysis
PyPDF2>=3.0.0
python-docx>=0.8.11
openpyxl>=3.1.0
# python-magic-bin>=0.4.14  # Not available on all platforms
pdfplumber>=0.9.0

# HTTP requests and web scraping
requests>=2.31.0
aiohttp>=3.8.0
beautifulsoup4>=4.12.0
feedparser>=6.0.0

# Environment and configuration
python-dotenv>=1.0.0

# Vector database and embeddings
chromadb>=0.4.0
sentence-transformers>=2.2.0

# Data validation and serialization
pydantic>=2.0.0

# Scheduling and background tasks
schedule>=1.2.0

# Date and time utilities
python-dateutil>=2.8.0

# Template engine
jinja2>=3.1.0

# File handling and utilities
# pathlib, io, base64, asyncio are built-in Python modules - no need to install